refactor(domain): 중복되는 역할을 하나의 객체에게 온전히 위임하였고 수반되는 변경사항들 반영
