+++ src/main/java/lotto/domain/Lottos.java
@@ -8,7 +8,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.StringJoiner;
-import lotto.domain.wrapper.PurchaseAmount;
 
 public class Lottos {
     private final List<Lotto> lottos;
