+++ src/main/java/lotto/domain/ProfitCalculator.java
@@ -1,7 +1,6 @@
 package lotto.domain;
 
 import java.util.Map;
-import lotto.domain.wrapper.PurchaseAmount;
 
 public class ProfitCalculator {
     private final Map<Prize, Integer> lottosResult;
@@ -10,11 +9,7 @@ public class ProfitCalculator {
         this.lottosResult = lottosResult;
     }
 
-    public double getProfitRate(PurchaseAmount purchaseAmount) {
-        return purchaseAmount.calculateProfitRate(getAllPrizeProfit());
-    }
-
-    private long getAllPrizeProfit() {
+    public long getTotalProfit() {
         long allPrizeProfit = 0;
         for (Prize prize : lottosResult.keySet()) {
             long prizeProfit = (long) prize.getPrizeProfit() * lottosResult.get(prize);
