+++ src/test/java/lotto/domain/wrapper/PurchaseAmountTest.java
@@ -8,6 +8,7 @@ import static lotto.utils.ErrorMessages.TOO_MUCH_PURCHASE_AMOUNT;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lotto.domain.PurchaseAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
